requirements.txt
# ============================================================
# AGENT ANONMUSK — Autonomous Bug Bounty Agent
# Dependencies
# ============================================================

# --- Core HTTP ---
requests>=2.31.0
httpx[http2]>=0.27.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=5.1.0

# --- LLM Providers ---
openai>=1.30.0
anthropic>=0.25.0

# --- Templating & Config ---
pyyaml>=6.0.1
jinja2>=3.1.3
python-dotenv>=1.0.1

# --- Reporting ---
markdown>=3.5.0

# --- CLI & UX ---
rich>=13.7.0
colorama>=0.4.6
tqdm>=4.66.0

# --- Data & Validation ---
pydantic>=2.6.0

# --- Task Queue (Optional — install for distributed mode) ---
# celery>=5.3.0
# redis>=5.0.0

# --- Testing ---
pytest>=8.0.0
pytest-asyncio>=0.23.0
